+++ CHANGELOG.md
@@ -6,6 +6,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+- Added ability to support empty variables [#124](https://github.com/cyberark/summon/issues/124)
+
+### Added
+- Added better errors for unkown tags found in the yaml
+
 ## [v0.8.0](https://github.com/cyberark/summon/releases/tag/v0.8.0) - 2019-09-20
 
 ### Changed