+++ CHANGELOG.md
@@ -5,11 +5,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+
+## [0.8.1] - 2020-03-02
 ### Changed
 - Added ability to support empty variables [#124](https://github.com/cyberark/summon/issues/124)
 
 ### Added
-- Added better errors for unkown tags found in the yaml
+- Added better errors for unknown tags found in the yaml
 - Added ability to set a default variable value with `default='<value>'` tag [#38](https://github.com/cyberark/summon/issues/38)
 
 ## [0.8.0] - 2019-09-20
@@ -151,7 +153,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Initial release
 
-[Unreleased]: https://github.com/cyberark/summon/compare/v0.8.0...HEAD
+[Unreleased]: https://github.com/cyberark/summon/compare/v0.8.1...HEAD
+[0.8.1]: https://github.com/cyberark/summon/compare/v0.8.0...v0.8.1
 [0.8.0]: https://github.com/cyberark/summon/compare/v0.7.0...v0.8.0
 [0.7.0]: https://github.com/cyberark/summon/compare/v0.6.11...v0.7.0
 [0.6.11]: https://github.com/cyberark/summon/compare/v0.6.10...v0.6.11